Fixed CR comments

+++ pkg/network/tracer/connection/kprobe/tracer.go
@@ -550,7 +550,7 @@ func populateConnStats(stats *network.ConnectionStats, t *netebpf.ConnTuple, s *
 		IsAssured:        s.IsAssured(),
 	}
 
-	if s.Protocol < uint16(network.MaxProtocols) {
+	if s.Protocol < uint8(network.MaxProtocols) {
 		stats.Protocol = network.ProtocolType(s.Protocol)
 	} else {
 		log.Warnf("got protocol %d which is not recognized by the agent", s.Protocol)
